+++ tests/io_test.py
@@ -1,5 +1,6 @@
 import torch
-import torchaudio
+from torchcodec.decoders import AudioDecoder, AudioStreamMetadata
+
 from pyannote.core import Segment
 from torch import Tensor
 
@@ -9,8 +10,8 @@ from pyannote.audio.core.io import Audio
 def test_audio_resample():
     "Audio is correctly resampled when it isn't the correct sample rate"
     test_file = "tests/data/dev00.wav"
-    info = torchaudio.info(test_file)
-    old_sr = info.sample_rate
+    metadata: AudioStreamMetadata = AudioDecoder(test_file).metadata
+    old_sr = metadata.sample_rate
     loader = Audio(sample_rate=old_sr // 2, mono="downmix")
     wav, sr = loader(test_file)
     assert isinstance(wav, Tensor)